Add warden test configuration
Add a helper for acceptance tests that allow you to login on behalf of a
user. This remediates the need to go to the login page and fill out the
form for every acceptance test requiring login. This also has the
benefit of closely resembling what's going on under the hood.